[Refactor] BE Socket API 분리
